Fix compiler warning in monitor.cc
At server/src/monitor.cc:88:64:
warning: field precision specifier ‘.*’ expects argument of type ‘int’,
but argument 3 has type ‘size_t {aka long unsigned int}’

+++ server/src/monitor/monitor.cc
@@ -85,7 +85,7 @@ public:
           auto *handler = Cmd_control::get()->find_cmd_handler(cmd, cmd_len);
 
           if (!handler)
-            fprintf(f, "Unknown command %.*s\n", cmd_len, cmd);
+            fprintf(f, "Unknown command %.*s\n", static_cast<int>(cmd_len), cmd);
           else
             handler->usage(f);
         }
